+++ packages/ui-components/src/common/hooks/index.ts
@@ -2,3 +2,5 @@ export { default as usePromise } from "./usePromise";
 export * from "./metric-data";
 export * from "./useResizeObserver";
 export * from "./useSvgBBox";
+export * from "./useHoveredDate";
+export * from "./useHoveredPoint";
