+++ packages/ui-components/src/common/hooks/useHoveredPoint.ts
@@ -1,7 +1,10 @@
 import { useState } from "react";
 import isNumber from "lodash/isNumber";
 import { Timeseries, TimeseriesPoint } from "@actnowcoalition/metrics";
-import { ChartOverlayXYProps, HoveredPointInfo } from "./ChartOverlayXY";
+import {
+  ChartOverlayXYProps,
+  HoveredPointInfo,
+} from "../../components/ChartOverlayXY";
 
 /**
  * React hook that keeps track of the point being hovered in a chart. This hook
