+++ packages/ui-components/src/components/BarChart/BarChart.tsx
@@ -4,7 +4,7 @@ import { Group } from "@visx/group";
 import { Timeseries } from "@actnowcoalition/metrics";
 import { useTheme } from "@mui/material";
 
-export interface BarChartOwnProps {
+export interface BaseBarChartProps {
   /** Timeseries used to draw the bar chart */
   timeseries: Timeseries<number>;
 
@@ -18,7 +18,7 @@ export interface BarChartOwnProps {
   barWidth?: number;
 }
 
-export type BarChartProps = BarChartOwnProps & React.SVGProps<SVGRectElement>;
+export type BarChartProps = BaseBarChartProps & React.SVGProps<SVGRectElement>;
 
 /**
  * BarChart is a chart to represent a Timeseries as a series of bars, one bar
