+++ packages/ui-components/src/components/ChartOverlayX/index.ts
@@ -1,2 +1,1 @@
 export * from "./ChartOverlayX";
-export * from "./useHoveredDate";
