+++ packages/ui-components/src/components/ChartOverlayXY/index.ts
@@ -1,2 +1,1 @@
 export * from "./ChartOverlayXY";
-export * from "./useHoveredPoint";
